requirements.txt
aiosqlite==0.19.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.7.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.2
sniffio==1.3.1
tzdata==2025.2
tzlocal==5.3.1
python-dotenv
